build(installer): package native PTY runtimes
